Shopping page without content for backbay
Finished the shopping page without content for backbay. This can be used to make each shopping page. All that would need to be changed is content and links. I think that we should add a nav bar that will link back to both the backbay homepage and the main homepage